# openapi.yaml
openapi: 3.0.3
info:
  title: panorama-builder API
  description: |
    JSON API for compositing flat photos into a 360° panorama. Stores
    stations (camera setups), photos (with embedded pose), image
    measurements (reticle anchors on photos), and control points
    (cross-station landmarks with a user-seeded location estimate).

    Image measurements link to a control point; the CP is the latent
    real-world feature each observation describes.

    All resources use server-assigned 13-character base32 IDs.
    Authentication is out of scope — anyone with the URL has read+write
    access. Field-level validation lives in handler code; this spec
    documents shapes only.
  version: 1.0.0
servers:
  - url: /api
...
paths:
...
  /stations:
    post:
      tags: [stations]
      summary: Create a station
      operationId: createStation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStationRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Station' }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [stations]
      summary: List stations, optionally filtered by bbox
      operationId: listStations
      parameters:
        - name: bbox
          in: query
          description: |
            Bounding box as `minLng,minLat,maxLng,maxLat` (four
            comma-separated floats). When omitted, returns up to 1000
            stations ordered by created_at DESC.
          required: false
          schema:
            type: string
            pattern: '^-?\d+(\.\d+)?,-?\d+(\.\d+)?,-?\d+(\.\d+)?,-?\d+(\.\d+)?$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Station' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /stations/{id}:
    parameters:
      - $ref: '#/components/parameters/StationId'
    get:
      tags: [stations]
      summary: Get a station with its photos, measurements, and referenced control points
      operationId: getStation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HydratedStation' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [stations]
      summary: Partial-update a station
      description: |
        Partial update: only the keys present in the request body are
        written. Omitting a key preserves its current value; sending an
        explicit `null` clears nullable columns (e.g. `name`).
      operationId: updateStation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StationUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Station' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [stations]
      summary: Delete a station and cascade
      description: |
        Cascades to photos and image measurements in the database, and
        removes photo blob files from disk. Control points referenced
        only from the deleted station are NOT cascade-deleted (they
        live cross-station) — they survive without measurements until
        explicitly removed.
      operationId: deleteStation
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /stations/{id}/photos:
    parameters:
      - $ref: '#/components/parameters/StationId'
    post:
      tags: [photos]
      summary: Create a photo on a station (metadata only)
      description: |
        Creates a photo row with an empty blob slot. Upload the image bytes
        via `PUT /photos/{id}/blob` afterwards.
      operationId: createPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhotoPosePatch' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /photos/{id}:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
    get:
      tags: [photos]
      summary: Get photo metadata
      operationId: getPhoto
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [photos]
      summary: Partial-update photo pose / opacity / aspect / locks
      description: |
        Partial update: only the keys present in the request body are
        written. Omitting a key preserves its current value.
      operationId: updatePhoto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhotoUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [photos]
      summary: Delete a photo (and its blob)
      operationId: deletePhoto
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /photos/{id}/blob:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
    put:
      tags: [photos]
      summary: Upload the photo's image blob
      description: |
        Stores the uploaded bytes content-addressed at `blobs/<sha256>` and
        writes the resulting path into the photo's `blob_path`. Re-uploading
        different bytes produces a new hash; the previous bytes remain on
        disk and are still referenced by any prior commit's view of the row.
      operationId: uploadPhotoBlob
      requestBody:
        required: true
        content:
          'image/*':
            schema:
              type: string
              format: binary
      responses:
        '204': { description: Stored }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Payload too large (above MAX_BLOB_BYTES)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [photos]
      summary: Download the photo's image blob
      description: |
        Resolves the photo's `blob_path` (overlay-aware when `X-Session-Id`
        is supplied) and streams the bytes. Callers that already have the
        content hash should prefer `GET /blobs/{hash}`, which serves the
        bytes directly with immutable cache headers and works without a
        session lookup.
      operationId: getPhotoBlob
      responses:
        '200':
          description: OK
          content:
            'image/*':
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
  /blobs/{hash}:
    parameters:
      - name: hash
        in: path
        required: true
        schema: { type: string, pattern: '^[0-9a-f]{64}$' }
    get:
      tags: [photos]
      summary: Download an image blob by its content hash
      description: |
        Content-addressed blob fetch. Bytes are immutable; the response is
        marked `Cache-Control: public, max-age=31536000, immutable`.
        Content-Type is sniffed from the first 512 bytes.
      operationId: getBlob
      responses:
        '200':
          description: OK
          content:
            'image/*':
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
  /photos/{id}/image-measurements:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
    post:
      tags: [image-measurements]
      summary: Create an image measurement on a photo
      operationId: createImageMeasurement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageMeasurementPatch' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageMeasurement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /image-measurements/{id}:
    parameters:
      - $ref: '#/components/parameters/ImageMeasurementId'
    put:
      tags: [image-measurements]
      summary: Partial-update an image measurement's u/v or control point link
      description: |
        Partial update: only the keys present in the request body are
        written. Sending `control_point_id: null` unlinks the measurement
        from its control point.
      operationId: updateImageMeasurement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageMeasurementUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageMeasurement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [image-measurements]
      summary: Delete an image measurement
      operationId: deleteImageMeasurement
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /control-points:
    post:
      tags: [control-points]
      summary: Create a control point
      description: |
        Control points are global (cross-station). Initial est_lat /
        est_lng / est_alt are optional; a CP without estimates is
        a "latent" landmark whose location is unknown until the solver
        triangulates from observations.
      operationId: createControlPoint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ControlPointPatch' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlPoint' }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [control-points]
      summary: List control points, optionally filtered by bbox
      operationId: listControlPoints
      parameters:
        - name: bbox
          in: query
          description: |
            Bounding box as `minLng,minLat,maxLng,maxLat`. Filter is
            applied to the CPs' est_lat/est_lng — CPs with NULL
            estimates are excluded when bbox is set. Returns up to 1000.
          required: false
          schema:
            type: string
            pattern: '^-?\d+(\.\d+)?,-?\d+(\.\d+)?,-?\d+(\.\d+)?,-?\d+(\.\d+)?$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ControlPoint' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /control-points/{id}:
    parameters:
      - $ref: '#/components/parameters/ControlPointId'
    get:
      tags: [control-points]
      summary: Get a control point
      operationId: getControlPoint
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlPoint' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [control-points]
      summary: Update a control point's description / estimates
      operationId: updateControlPoint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ControlPointPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlPoint' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [control-points]
      summary: Delete a control point
      description: |
        Cascades `ON DELETE SET NULL` on referencing image and map
        measurements; both survive without their link.
      operationId: deleteControlPoint
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /control-point-constraints:
    post:
      tags: [cp-constraints]
      summary: Create a pairwise constraint between two control points
      description: |
        Creates a hard equality constraint between two control points.
        The handler swaps `cp_a_id` and `cp_b_id` if necessary so the
        stored row satisfies the canonical `cp_a_id < cp_b_id` ordering.
        Self-loops (same id on both sides) and duplicate (a, b, type)
        triples are rejected.

        Type semantics:
          * `plumb` — the two CPs share `est_lat` and `est_lng`
          * `level` — the two CPs share `est_alt`
      operationId: createCPConstraint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CPConstraintCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CPConstraint' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A constraint of this type already exists between this pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [cp-constraints]
      summary: List all pairwise control-point constraints
      operationId: listCPConstraints
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CPConstraint' }
  /control-point-constraints/{id}:
    parameters:
      - $ref: '#/components/parameters/CPConstraintId'
    put:
      tags: [cp-constraints]
      summary: Change the constraint type
      operationId: updateCPConstraint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CPConstraintPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CPConstraint' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [cp-constraints]
      summary: Delete a constraint
      operationId: deleteCPConstraint
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /control-point-surfaces:
    post:
      tags: [cp-surfaces]
      summary: Create a polygonal surface anchored on 3 or 4 control points
      description: |
        Creates a triangle (3 distinct CP ids) or quad (4 distinct CP ids) surface
        for visualization in the 3D world view. The four CP ids must be pairwise
        distinct; `cp_4_id` omitted or null yields a triangle. Surfaces have no
        editable fields after creation — modify by deleting and recreating.
      operationId: createCPSurface
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CPSurfaceCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CPSurface' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      tags: [cp-surfaces]
      summary: List all CP surfaces
      operationId: listCPSurfaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CPSurface' }
  /control-point-surfaces/{id}:
    parameters:
      - $ref: '#/components/parameters/CPSurfaceId'
    delete:
      tags: [cp-surfaces]
      summary: Delete a surface
      operationId: deleteCPSurface
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /solve/joint:
    post:
      tags: [solver]
      summary: Bundle-adjust the entire system (stations + photos + control points)
      description: |
        Runs Gauss-Newton over every unlocked solvable parameter across all
        stations, photos, and control points. Requires at least two stations
        with both lat and lng locked (translation + rotation gauge); returns
        400 otherwise. Persists the result inside one transaction; uses
        optimistic concurrency on each row's `updated_at` and returns 409 if
        another writer raced the solve.
      operationId: solveJoint
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SolveConfig' }
      responses:
        '200':
          description: OK (incl. divergence; check `diverged`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Concurrent edit detected; re-fetch and retry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /solve/stations/{id}:
    parameters:
      - $ref: '#/components/parameters/StationId'
    post:
      tags: [solver]
      summary: Refine the photo poses of a single station
      description: |
        Solves the joint photo-pose problem for the given station, treating
        the station's own lat/lng and every control point's est_* as fixed.
        No gauge requirement (the station is the implicit anchor).
      operationId: solveStation
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SolveConfig' }
      responses:
        '200':
          description: OK (incl. divergence; check `diverged`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Concurrent edit detected; re-fetch and retry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /solve/control-points/{id}:
    parameters:
      - $ref: '#/components/parameters/ControlPointId'
    post:
      tags: [solver]
      summary: Refine est_lat / est_lng / est_alt for a single control point
      description: |
        Triangulates the CP location from its image measurements across
        stations. Requires at least two image observations.
      operationId: solveControlPoint
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SolveConfig' }
      responses:
        '200':
          description: OK (incl. divergence; check `diverged`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Concurrent edit detected; re-fetch and retry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /control-points/{id}/observations:
    parameters:
      - $ref: '#/components/parameters/ControlPointId'
    get:
      tags: [control-points]
      summary: List every observation referencing this control point
      description: |
        Returns image measurements across every station that link back
        to the given control point. Each row carries enough context
        (station id + name) for the caller to deep-link.
      operationId: listControlPointObservations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlPointObservations' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    get:
      tags: [sessions]
      summary: Session metadata + ops summary + live conflict list
      operationId: getSession
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionState' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sessions/{id}/ops:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    get:
      tags: [sessions]
      summary: Full journal of ops in seq order
      operationId: getSessionOps
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SessionOp' }
  /sessions/{id}/abandon:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    post:
      tags: [sessions]
      summary: Mark a session as abandoned
      description: |
        Idempotent: returns 204 if already abandoned. Refuses with 409 if
        the session is merged (history is immutable).
      operationId: abandonSession
      responses:
        '204': { description: Abandoned }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Session is already merged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /sessions/{id}/merge:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    post:
      tags: [sessions]
      summary: Promote the session to a new commit on main
      description: |
        Coalesces session ops per entity, applies them to main inside one
        SERIALIZABLE transaction, and writes a new commit row referencing
        this session. Refuses with 409 if any entity touched by the
        session has been modified by a commit since the session's base.
      operationId: mergeSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MergeRequest' }
      responses:
        '201':
          description: Merged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommitRef' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Conflict — abandon and start over
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictsResponse' }
  /commits:
    get:
      tags: [commits]
      summary: List commits, latest first
      operationId: listCommits
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500 }
        - name: before_seq
          in: query
          required: false
          description: Return commits whose seq is strictly less than this value (cursor pagination).
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Commit' }
  /commits/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    get:
      tags: [commits]
      summary: One commit row plus the journaled ops it landed
      operationId: getCommit
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommitWithOps' }
        '404': { $ref: '#/components/responses/NotFound' }
  /commits/{id}/revert:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    post:
      tags: [commits]
      summary: Create a new commit that inverts the target's effect
      description: |
        Walks the target commit's journal backward, swaps before/after on
        each op, and applies the inverse plan to main. Refuses with 409 if
        any later commit has touched any of the target's entities.
      operationId: revertCommit
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommitMessage' }
      responses:
        '201':
          description: Reverted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommitRef' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Conflict — later commits have touched these entities
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictsResponse' }
  /control-points/{id}/visible-photos:
    parameters:
      - $ref: '#/components/parameters/ControlPointId'
    get:
      tags: [control-points]
      summary: List photos whose viewshed should contain this control point
      description: |
        Returns photos where the CP's estimated location falls inside the
        photo's horizontal field of view AND the station's capture time
        falls inside the CP's lifespan AND no image measurement for this
        CP exists yet on the photo. The result is the inverse of the
        observations list — candidate photos still missing a measurement.

        Returns an empty list (200) when the CP has no estimated
        location; the caller decides how to render that state.
      operationId: listControlPointVisiblePhotos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControlPointVisiblePhotos' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Station:
      type: object
      required: [id, lat, lng, alt, name, lock_lat, lock_lng, lock_alt, captured_at, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lng: { type: number, format: double, minimum: -180, maximum: 180 }
        alt: { type: number, format: double, description: meters above sea level (defaults to 0 when unknown) }
        name: { type: string, nullable: true }
        lock_lat: { type: boolean, description: when true the solver leaves lat untouched }
        lock_lng: { type: boolean, description: when true the solver leaves lng untouched }
        lock_alt: { type: boolean, description: when true the solver treats alt as a fixed input }
        captured_at:
          type: string
          format: date-time
          description: when the photographer's camera was set up at this site
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CreateStationRequest:
      type: object
      description: |
        POST body. lat/lng/captured_at are required; other fields default if omitted.
      required: [lat, lng, captured_at]
      properties:
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lng: { type: number, format: double, minimum: -180, maximum: 180 }
        alt: { type: number, format: double, description: defaults to 0 when omitted }
        name: { type: string, nullable: true }
        lock_lat: { type: boolean, description: defaults to false when omitted }
        lock_lng: { type: boolean }
        lock_alt: { type: boolean }
        captured_at: { type: string, format: date-time }
...